Transform `react-colorful` when running tests
`react-colorful` ships as ES Modules.  Support for modules is still experimental in Node and Jest, so we need to ensure the source for this package is transformed to CJS modules when running the tests.

The default Jest config doesn’t transform any files inside of `node_modules` — this ensure that it does make an exception for `react-colorful` next to the other packages already in the list.