Fix color picker translations
When I added labels for the color picker swatches, I did not wrap the component in an IntlProvider (using withTranslator).

When used in the context of network diagrams, this doesn’t cause bugs, as the network diagram itself is wrapped in an IntlProvider. The bug only surfaced once we upgraded react-ftm in Aleph, as Aleph also uses the color picker component in the timelines UI.